split diferent data stores to different files
