+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/Settings/Settings.cs
@@ -93,7 +93,7 @@ namespace DCL.SettingsCommon
                 mouseSensitivity = 0.6f,
                 scenesLoadRadius = 4,
                 avatarsLODDistance = 16,
-                maxNonLODAvatars = DataStore.DataStore_AvatarsLOD.DEFAULT_MAX_AVATAR,
+                maxNonLODAvatars = DataStore_AvatarsLOD.DEFAULT_MAX_AVATAR,
                 voiceChatVolume = 1,
                 voiceChatAllow = GeneralSettings.VoiceChatAllow.ALL_USERS,
                 autoqualityOn = false,
