+++ unity-renderer/Assets/Scripts/MainScripts/DCL/DataStore/DataStore_Camera.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace DCL
+{
+    public class DataStore_Camera
+    {
+        public readonly BaseVariable<Quaternion> rotation =  new BaseVariable<Quaternion>();
+    }
+}
