+++ unity-renderer/Assets/Scripts/MainScripts/DCL/DataStore/DataStore_FeatureFlag.cs
@@ -1,0 +1,7 @@
+namespace DCL
+{
+    public class DataStore_FeatureFlag
+    {
+        public readonly BaseVariable<FeatureFlag> flags = new BaseVariable<FeatureFlag>(new FeatureFlag());
+    }
+}
