+++ unity-renderer/Assets/Scripts/MainScripts/DCL/DataStore/DataStore_Quests.cs
@@ -1,0 +1,8 @@
+namespace DCL
+{
+    public class DataStore_Quests
+    {
+        public readonly BaseDictionary<string, QuestModel> quests = new BaseDictionary<string, QuestModel>();
+        public readonly BaseCollection<string> pinnedQuests = new BaseCollection<string>();
+    }
+}
