+++ unity-renderer/Assets/Scripts/MainScripts/DCL/DataStore/DataStore_Screen.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace DCL
+{
+    public class DataStore_Screen
+    {
+        public readonly BaseVariable<Vector2Int> size = new BaseVariable<Vector2Int>(Vector2Int.zero);
+    }
+}
