+++ unity-renderer/Assets/Scripts/MainScripts/DCL/DataStore/DataStore_Settings.cs
@@ -1,0 +1,7 @@
+namespace DCL
+{
+    public class DataStore_Settings
+    {
+        public readonly BaseVariable<bool> profanityChatFilteringEnabled = new BaseVariable<bool>();
+    }
+}
